convert to es6 module
