+++ .gitignore
@@ -1,4 +1,5 @@
 node_modules/
+dist/
 package-lock.json
 /pkg
 /pkg.tgz
