+++ tsconfig.json
@@ -1,0 +1,15 @@
+{
+	"extends": "@companion-module/tools/tsconfig/node18/recommended",
+	"include": ["./**/*.ts"],
+	"exclude": ["node_modules/**", "./**/*spec.ts", "./**/__tests__/*", "./**/__mocks__/*"],
+	"compilerOptions": {
+		"outDir": "./dist",
+		"baseUrl": "./",
+		"paths": {
+			"*": ["./node_modules/*"]
+		},
+		"noImplicitAny": false,
+		"module": "ES6",
+		"esModuleInterop": true,
+	}
+}
