--- upgrades.js
@@ -1,9 +1,0 @@
-module.exports = {
-	function (context, props) {
-		return {
-			updatedConfig: null,
-			updatedActions: [],
-			updatedFeedbacks: [],
-		}
-	},
-}
